+++ src/algoritmia/examples/backtracking/hamiltonian_cycle_bt.py
@@ -6,40 +6,44 @@ from algoritmia.schemes.bt_scheme import DecisionSequence, bt_solutions, min_sol
 
 # Tipos  --------------------------------------------------------------------------
 
-# En este problema, una decisión es un vértice, por lo tanto, su tipo es T (el genérico del grafo)
-# Decision = T
+# En este problema, una decisión es un vértice, por lo tanto, su tipo es TVertex (el genérico del grafo)
+# TDecision = TVertex
 
-# 'bt_solutions' y 'bt_vc_solutions' devuelven un Iterator del tipo devuelto por el método 'solution' de
-# la clase 'DecisionSequence', cuya implementación por defecto devuelve una tupla con las decisiones:
-type Solution[T] = tuple[T, ...]
+# Queremos que una solución sea la secuencia de decisiones (vértices) en forma de tupla:
+type Solution[TDecision] = tuple[TDecision, ...]
 
+# - 'bt_solutions' y 'bt_vc_solutions' devuelven un Iterator con las DecisionSequence que
+#   llegan a una solución.
+# - Pero un objeto DecisionSequence no es una tupla de decisiones: debemos utilizar el método
+#   'decisions()' de la clase DecisionSequence para obtener la tupla.
 
 # --------------------------------------------------------------------------------
 
-def hamiltoniancycle_solutions[T](g: IGraph[T]) -> Iterator[Solution]:
-    class HamiltonianCycleDS(DecisionSequence[T, None]):  # Como no hay Extra -> ponemos None
+def hamiltoniancycle_solutions[TVertex](g: IGraph[TVertex]) -> Iterator[Solution[TVertex]]:
+    class HamiltonianCycleDS(DecisionSequence[TVertex, None]):  # Como no hay Extra -> ponemos None
         def is_solution(self) -> bool:
-            return len(self) == len(g.V) and v_initial in g.succs(self.decision)
+            return len(self) == len(g.V) and v_initial in g.succs(self.last_decision())
 
         def successors(self) -> Iterator[Self]:
             if len(self) < len(g.V):
                 ds_set = set(self.decisions())  # O(|V|)
-                for v in g.succs(self.decision):
+                for v in g.succs(self.last_decision()):
                     if v not in ds_set:  # O(1), |V| veces
                         yield self.add_decision(v)
 
     initial_ds = HamiltonianCycleDS()
     v_initial = next(iter(g.V))  # Elegimos un vértice cualquiera como inicial
     initial_ds2 = initial_ds.add_decision(v_initial)  # Forzamos la primera decisión
-    return bt_solutions(initial_ds2)
+    for ds_sol in bt_solutions(initial_ds2):
+        yield ds_sol.decisions()  # Extraemos las decisiones del objeto ds_sol y las devolvemos
 
 
-Score = int | float  # La longitud (suma de pesos) del ciclo
-ScoredSolution = tuple[Score, Solution]
+type Score = int | float    # La longitud (suma de pesos) del ciclo
+type ScoredSolution[TDecision] = tuple[Score, Solution[TDecision]]
 
 
-def hamiltoniancycle_best_solution[T](g: IGraph[T],
-                                      wf: WeightingFunction[T]) -> Optional[ScoredSolution]:
+def hamiltoniancycle_best_solution[TVertex](g: IGraph[TVertex],
+                                            wf: WeightingFunction[TVertex]) -> Optional[ScoredSolution[TVertex]]:
     def f(sol: Solution) -> Score:
         return sum(wf(sol[i - 1], sol[i]) for i in range(len(sol)))
 
@@ -51,9 +55,9 @@ def hamiltoniancycle_best_solution[T](g: IGraph[T],
 if __name__ == "__main__":
     edges = [(0, 2), (0, 3), (0, 9), (1, 3), (1, 4), (1, 8), (2, 3), (2, 5), (3, 4),
              (3, 6), (4, 7), (5, 6), (5, 8), (6, 7), (6, 8), (6, 9)]
-    g0 = UndirectedGraph(E=edges)
+    g0 = UndirectedGraph[int](E=edges)
     d0 = dict(((u2, v2), abs(u2 - v2)) for (u2, v2) in g0.E)
-    wf0 = WeightingFunction(d0, symmetrical=True)
+    wf0 = WeightingFunction[int](d0, symmetrical=True)
 
     print('Basic versión (all solutions):')
     has_solutions = False