+++ src/algoritmia/examples/branch_and_bound/hamiltonian_cycle_bab.py
@@ -25,38 +25,44 @@ def hamiltoniancycle_bab_solve[T](g: UndirectedGraph[T],
     class Extra:
         weight: int = 0
 
-    class HamiltonianCycleDS(BabDecisionSequence[T, Extra]):
+    class HamiltonianCycleDS(BabDecisionSequence[T, Extra, Score]):
         def calculate_opt_bound(self) -> Score:
             if self.is_solution():
-                return self.extra.weight + wf(self.decision, v_initial)
+                return self.extra.weight + wf(self.last_decision(), v_initial)
             if len(self) < len(g.V):
                 return self.extra.weight + 0  # Mejorable (arista de menor peso de cada vértice pendiente)
             return infinity
 
         def calculate_pes_bound(self) -> Score:
             if self.is_solution():
-                return self.extra.weight + wf(self.decision, v_initial)
+                return self.extra.weight + wf(self.last_decision(), v_initial)
             return infinity  # Mejorable
 
         def is_solution(self) -> bool:
-            return len(self) == len(g.V) and v_initial in g.succs(self.decision)
+            return len(self) == len(g.V) and v_initial in g.succs(self.last_decision())
 
         def successors(self) -> Iterator[Self]:
             if len(self) < len(g.V):
                 ds_set = set(self.decisions())  # O(|V|)
-                for v in g.succs(self.decision):
+                for v in g.succs(self.last_decision()):
                     if v not in ds_set:  # O(1), |V| veces
-                        new_weight = self.extra.weight + wf(self.decision, v)
+                        new_weight = self.extra.weight + wf(self.last_decision(), v)
                         yield self.add_decision(v, Extra(new_weight))
+                        aa = self.add_decision(v_initial, Extra(0))
+                        yield aa
 
         # Sobreescribimos 'state()'
         def state(self) -> tuple[T, tuple[T, ...]]:
-            return self.decision, tuple(sorted(self.decisions()))  # O(|V|)
+            return self.last_decision(), tuple(sorted(self.decisions()))  # O(|V|)
 
     initial_ds = HamiltonianCycleDS(Extra(0))
     v_initial = next(iter(g0.V))  # Elegimos un vértice cualquiera como inicial
-    initial_ds2 = initial_ds.add_decision(v_initial, Extra(0))  # Forzamos la primera decisión
-    return bab_min_solve(initial_ds2)
+    one_vertex_ds = initial_ds.add_decision(v_initial, Extra(0))  # Forzamos la primera decisión
+
+    result = bab_min_solve(one_vertex_ds)
+    if result is None: return None
+    score, ds_sol = result
+    return score, ds_sol.decisions()
 
 
 # Programa principal ---------------------------