+++ src/algoritmia/examples/dynamic_programming/edit_distance.py
@@ -4,7 +4,7 @@ type Decision = str
 type Solution = list[Decision]
 
 type Score = int
-type ScoredSolution = tuple[Score, Solution]
+type Result = tuple[Score, Solution]  # Siempre habrá solución, no hace falta añadir None
 
 type SParams = tuple[int, int]  # (m, n)
 
@@ -14,10 +14,11 @@ class RecMode(Enum):
     Optimized = 1
 
 
-def edit_distance(mode: RecMode, s: str, t: str) -> ScoredSolution:
+def edit_distance(mode: RecMode, s: str, t: str) -> Result:
     def D_classic(m: int, n: int) -> Score:
-        if m == 0 and n == 0:
-            return 0
+        # Caso base
+        if m == 0 and n == 0: return 0
+        # Recursividad memoizada
         if (m, n) not in mem:
             if n == 0:
                 mem[m, n] = D_classic(m - 1, n) + 1, (m - 1, n), 'D'
@@ -31,8 +32,9 @@ def edit_distance(mode: RecMode, s: str, t: str) -> ScoredSolution:
         return mem[m, n][0]
 
     def D_optimized(m: int, n: int) -> Score:
-        if m == 0 and n == 0:
-            return 0
+        # Caso base
+        if m == 0 and n == 0: return 0
+        # Recursividad memoizada
         if (m, n) not in mem:
             if n == 0:
                 mem[m, n] = D_optimized(m - 1, n) + 1, (m - 1, n), 'D'
@@ -53,14 +55,15 @@ def edit_distance(mode: RecMode, s: str, t: str) -> ScoredSolution:
     else:
         score = D_optimized(len(s), len(t))
 
-    sol = []
+    # Recuperamos la solución de mem
+    decisions = []
     m, n = len(s), len(t)
     while (m, n) != (0, 0):
         _, (m, n), dec = mem[m, n]
-        sol.append(dec)
-    sol.reverse()
+        decisions.append(dec)
+    decisions.reverse()
 
-    return score, sol
+    return score, decisions
 
 
 # ------------------------------------------------------------------------------------------------------
@@ -70,5 +73,5 @@ if __name__ == '__main__':
         my_source0, my_target0 = 'costa', 'casa'
         print(f'Edit distance fom "{my_source0}" to "{my_target0}":')
         print(f'\tMode: {mode0}')
-        res_score0, res_decisions0 = edit_distance(mode0, my_source0, my_target0)
-        print(f'\tScore: {res_score0}\n\tDecisions: {res_decisions0}\n')
+        result0 = edit_distance(mode0, my_source0, my_target0)
+        print(f'\tResult: {result0}\n')