+++ src/algoritmia/schemes/dac_scheme.py
@@ -1,37 +1,35 @@
 """
 
-Version: 5.2 (01-dic-2023)
+Version: 6.0 (23-oct-2024) - Añade tipo genérico TSolution y cambia Iterator por Iterable
+         5.2 (01-dic-2023)
          4.0 (23-oct-2021)
 
 @author: David Llorens (dllorens@uji.es)
-         (c) Universitat Jaume I 2021
+         (c) Universitat Jaume I 2024
 @license: GPL3
 """
 from abc import ABC, abstractmethod
-from collections.abc import Iterator
-from typing import Any, Self
-
-type Solution = Any
+from collections.abc import Iterable
+from typing import Self
 
 
 # Divide and conquer -----------------------------------------------------------
 
-
-class IDivideAndConquerProblem(ABC):
+class IDivideAndConquerProblem[TSolution](ABC):
     @abstractmethod
     def is_simple(self) -> bool: pass
 
     @abstractmethod
-    def trivial_solution(self) -> Solution: pass
+    def trivial_solution(self) -> TSolution: pass
 
     @abstractmethod
-    def divide(self) -> Iterator[Self]: pass
+    def divide(self) -> Iterable[Self]: pass
 
     @abstractmethod
-    def combine(self, solutions: Iterator[Solution]) -> Solution: pass
+    def combine(self, solutions: Iterable[TSolution]) -> TSolution: pass
 
 
-def div_solve(problem: IDivideAndConquerProblem) -> Solution:
+def div_solve[TSolution](problem: IDivideAndConquerProblem[TSolution]) -> TSolution:
     if problem.is_simple():
         return problem.trivial_solution()
     else:
@@ -42,21 +40,21 @@ def div_solve(problem: IDivideAndConquerProblem) -> Solution:
 
 # Decrease and conquer -----------------------------------------------------------
 
-class IDecreaseAndConquerProblem(ABC):
+class IDecreaseAndConquerProblem[TSolution](ABC):
     @abstractmethod
     def is_simple(self) -> bool: pass
 
     @abstractmethod
-    def trivial_solution(self) -> Solution: pass
+    def trivial_solution(self) -> TSolution: pass
 
     @abstractmethod
     def decrease(self) -> Self: pass
 
-    def process(self, s: Solution) -> Solution:
+    def process(self, s: TSolution) -> TSolution:
         return s
 
 
-def dec_solve(problem: IDecreaseAndConquerProblem) -> Solution:
+def dec_solve[TSolution](problem: IDecreaseAndConquerProblem[TSolution]) -> TSolution:
     if problem.is_simple():
         return problem.trivial_solution()
     else:
@@ -64,14 +62,14 @@ def dec_solve(problem: IDecreaseAndConquerProblem) -> Solution:
         return problem.process(dec_solve(smaller_problem))
 
 
-def tail_dec_solve(problem: IDecreaseAndConquerProblem) -> Solution:
+def tail_dec_solve[TSolution](problem: IDecreaseAndConquerProblem[TSolution]) -> TSolution:
     if problem.is_simple():
         return problem.trivial_solution()
     else:
         return tail_dec_solve(problem.decrease())
 
 
-def iter_dec_solve(problem: IDecreaseAndConquerProblem) -> Solution:
+def iter_dec_solve[TSolution](problem: IDecreaseAndConquerProblem[TSolution]) -> TSolution:
     while not problem.is_simple():
         problem = problem.decrease()
     return problem.trivial_solution()